requirements.txt
numpy
pyqtgraph
pyserial
PyQt5
